# study/testing/apifox/openapi/level2_04_service_layer.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "实现服务层",
    "description": "演示真正的三层架构：传输层 → 服务层 → 基础设施层",
    "version": "2.0.0"
  },
  "paths": {
    "/api/users": {
      "get": {
        "summary": "List Users",
        "description": "列出所有用户",
        "operationId": "list_users_api_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response List Users Api Users Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create User",
        "description": "创建用户\n\n══════════════════════════════════════════════════════════════════════════\n传输层职责（只有这些）：\n══════════════════════════════════════════════════════════════════════════\n\n✅ 接收请求\n✅ 参数校验（Pydantic）\n✅ 调用服务层\n✅ 返回响应\n✅ 异常处理（将领域异常转为 HTTP 响应）\n\n❌ 不包含业务逻辑\n❌ 不直接操作数据库\n❌ 不包含业务规则\n\n══════════════════════════════════════════════════════════════════════════",
        "operationId": "create_user_api_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "summary": "Get User",
        "description": "获取用户",
        "operationId": "get_user_api_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete User",
        "description": "删除用户",
        "operationId": "delete_user_api_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/email": {
      "put": {
        "summary": "Update User Email",
        "description": "更新用户邮箱",
        "operationId": "update_user_email_api_users__user_id__email_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "new_email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "title": "New Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/architecture/comparison": {
      "get": {
        "summary": "Compare Architectures",
        "description": "架构对比总结\n\n══════════════════════════════════════════════════════════════════════════\nLevel 1: 没有分层（问题）\n══════════════════════════════════════════════════════════════════════════\n\n@app.post(\"/users\")\nasync def create_user(user_data: UserCreate):\n    # ❌ 业务逻辑在 endpoint 中\n    if await db.query(\"SELECT * FROM users WHERE email = ?\", user_data.email):\n        raise HTTPException(400, \"Email exists\")\n\n    # ❌ 直接操作数据库\n    hashed = hash_password(user_data.password)\n    user_id = await db.insert(\"INSERT INTO users ...\")\n\n    return {\"id\": user_id}\n\n问题：\n- Endpoint 包含业务逻辑\n- 无法复用（CLI、gRPC 都用不了）\n- 难以测试（必须启动 HTTP）\n- 代码重复（多个 endpoint 写类似逻辑）\n\n══════════════════════════════════════════════════════════════════════════\nLevel 2: 三层架构（解决）\n══════════════════════════════════════════════════════════════════════════\n\n# Endpoint（传输层）\n@app.post(\"/users\")\nasync def create_user(\n    user_data: UserCreate,\n    service: UserService = Depends(get_user_service)\n):\n    # ✅ 只做协议适配\n    return await service.create_user(...)\n\n# Service（服务层）\nclass UserService:\n    async def create_user(self, ...):\n        # ✅ 业务逻辑在这里\n        if await self.repo.exists_by_email(email):\n            raise UserDuplicateError()\n        user = User.create(...)\n        return await self.repo.save(user)\n\n# Repository（基础设施层）\nclass InMemoryUserRepository:\n    async def save(self, user: User):\n        # ✅ 数据持久化\n        ...\n\n优势：\n- Endpoint 只做协议适配\n- 业务逻辑可复用、可测试\n- 代码清晰、职责分明\n- 易于维护和扩展\n\n══════════════════════════════════════════════════════════════════════════",
        "operationId": "compare_architectures_architecture_comparison_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "创建用户请求模型"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "用户响应模型"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}